x/stellartalk/types/message_create_chat_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"stellartalk/testutil/sample"
)

func TestMsgCreateChat_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateChat
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateChat{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateChat{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
